完成联系人除去 websocket 的部分

+++ api/base.go
@@ -2,7 +2,7 @@ package api
 
 // BaseRes 所有回复的基类
 type BaseRes struct {
-	Code    int         `json:"code"`
+	Code    int64       `json:"code"`
 	Msg     string      `json:"msg"`
 	Payload interface{} `json:"payload"`
 }
