+++ db/base/model.go
@@ -8,7 +8,7 @@ import (
 
 // Model 数据库 dao 类型的基类
 type Model struct {
-	ID        int `gorm:"primarykey" json:"id"`
+	ID        int64 `gorm:"primarykey" json:"id"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	DeletedAt gorm.DeletedAt `gorm:"index"`
